requirements.txt
docxtpl==0.4.7
python-docx==0.8.6
